Apply suggestions from code review

+++ src/entity/player_stats.hpp
@@ -24,8 +24,8 @@ public:
     GETTER_SETTER(float, CritChance, crit_chance)
     GETTER_SETTER(float, CritMultiplier, crit_multiplier)
     GETTER_SETTER(float, LoadingTime, loadingTime)
-    GETTER_SETTER(float, Dash, dash)
-    GETTER_SETTER(float, Pierce, pierce)
+    GETTER_SETTER(int, Dash, dash)
+    GETTER_SETTER(int, Pierce, pierce)
     GETTER_SETTER(float, Sizeof_area, sizeof_area)
     GETTER_SETTER(float, Size, size)
